rag/requirements.txt
fastapi
uvicorn
psycopg2-binary
langchain
langchain-google-genai
langchain-postgres
sqlalchemy
pydantic
python-dotenv
google-genai
aio_pika
